refactor: Refactor otp class.
